Part 3: only correlate complete observations
I thought this was what was already happening, since we filtered out things
that didn't meet the threshold.  But I was getting a bunch of NAs where I was
supposed to get numbers, and so I thought that looked like a reasonable value
to try for `use`.  And whatdya know, it matched the output.